+++ index.js
@@ -21,11 +21,8 @@ const contentSecurityPolicy = require("helmet-csp");
 const morgan = require("morgan");
 const rfs = require("rotating-file-stream");
 
-const router = require("./app/routes/router");
-const authRouter = require("./app/routes/auth");
-const usersRouter = require("./app/routes/users");
-const clientsRouter = require("./app/routes/clients");
-const api = require("./app/api/endpoints");
+const router = require("./app/router");
+const api = require("./api");
 
 const { errors, pageNotFound } = require("./app/middlewares/errors");
 
@@ -120,9 +117,6 @@ app.use(layout);
 
 // Rutas
 app.use("/", router);
-app.use("/", authRouter);
-app.use("/", usersRouter);
-app.use("/", clientsRouter);
 
 // API
 app.use("/api", api);
