+++ views/clients.ejs
@@ -15,10 +15,10 @@
     ></table>
   </div>
 </div>
-<%- include('partials/errorRowModal.ejs') %> <%-
-include('partials/formularios/clients/newClientForm') %> <%-
-include('partials/formularios/clients/editClientForm') %> <%-
-contentFor('styles') %>
+<%- include('partials/errorRowModal.ejs') %> 
+<%- include('partials/formularios/clients/newClientForm') %> 
+<%- include('partials/formularios/clients/editClientForm') %> 
+<%- contentFor('styles') %>
 <!--Datatables-->
 <link
   rel="stylesheet"
