+++ views/tickets.ejs
@@ -10,7 +10,7 @@
   </div>
   <div class="card-body">
     <table
-      id="clientsDataTable"
+      id="ticketsDataTable"
       class="display responsive table table-hover wrap w-100 fs-6"
     ></table>
   </div>
@@ -44,3 +44,66 @@
 <script src="https://cdn.datatables.net/1.13.6/js/dataTables.bootstrap5.min.js"></script>
 <script src="https://cdn.datatables.net/buttons/2.4.2/js/dataTables.buttons.min.js"></script>
 <script src="https://cdn.datatables.net/select/1.7.0/js/dataTables.select.min.js"></script>
+<script nonce="<%=cspNonce%>">
+  const table = $("#ticketsDataTable").DataTable({
+    responsive: true,
+    select: {
+      info: false,
+      style: "single",
+    },
+    processing: true,
+    serverSide: true,
+    ajax: "/api/clients", // Ruta de la API en tu servidor Express.js
+    columns: [
+      { title: "ID", data: "id" },
+      { title: "NOMBRE", data: "name" },
+      { title: "TELÉFONO", data: "phone" },
+      { title: "CORREO", data: "email" },
+      { title: "DIRECCIÓN", data: "address", width: "30%" }, // Ajusta el ancho de la columna de dirección al 30% del contenedor
+      { title: "PROVINCIA", data: "province_name" },
+      {
+        title: "ESTADO",
+        data: "status",
+        render: function (data) {
+          if (data === "HABILITADO") {
+            return `
+              <button title="INHABILITADO" class="btn btn-success btn-sm w-100" style="pointer-events:none">Habilitado</button>`;
+          } else {
+            return `
+              <button title="INHABILITADO" class="btn btn-danger btn-sm w-100 nowrap" style="pointer-events:none">Inhabilitado</button>`;
+          }
+        },
+        width: "11%",
+      },
+    ],
+    oLanguage: { sSearch: "" },
+    language: {
+      url: "https://cdn.datatables.net/plug-ins/1.13.6/i18n/es-ES.json",
+      searchPlaceholder: "Buscar...",
+    },
+    dom: '<"d-sm-flex justify-content-sm-between"lf><"my-1 d-flex justify-content-center" B>rt<"d-flex justify-content-center" i><"d-flex justify-content-center"p>',
+    buttons: [
+      {
+        text: '<i class="fas fa-plus"></i> Crear',
+        action: function (e, dt, node, config) {
+          $("#crearModal").modal("show");
+        },
+        className: "btn btn-success btn-sm",
+      },
+      {
+        text: '<i class="fas fa-pencil-alt"></i> Editar',
+        action: function (e, dt, node, config) {
+          // Obtener la fila seleccionada
+          const data = table.row({ selected: true }).data();
+          const editForm = $("#editForm");
+          if (data) {
+            showEditModal(editForm, data);
+          } else {
+            $("#rowErrorModal").modal("show");
+          }
+        },
+        className: "btn btn-warning btn-sm",
+      },
+    ],
+  });
+</script>
